feat(workflows): add Tinybird CI/CD workflows
Add Tinybird CI/CD workflows to the repository. The `tinybird_cd.yml` file sets up the CD workflow, which deploys changes to the 'unkey' workspace. The `tinybird_ci.yml` file sets up the CI workflow, which runs on pull requests to the 'main' branch. Both workflows use the `tinybirdco/ci` action with the specified version. The workflows require the `TINYBIRD_ADMIN_TOKEN` secret to be set in GitHub secrets and use the `https://api.tinybird.co` as the Tinybird host.