fix: tinybird syntax (#636)
* feat: add materialized views

* chore(mv__monthly_verifications.datasource): revert to VERSION 1 and update schema

The change was made to revert the datasource to VERSION 1 and update the schema. The previous version (VERSION 2) was created from a Pipe 'aggregate_verifications_daily__v2', but it was decided to revert to the previous version and update the schema accordingly.

* fix(aggregate_verifications_monthly_by_workspace.pipe): remove trailing comma in SQL query

The trailing comma in the SQL query was causing a syntax error. This commit removes the trailing comma to fix the issue.