feat(aggregate_verifications_monthly_by_workspace.pipe): add time column to the aggregation query
The change adds a new column called "time" to the aggregation query in the `aggregate_verifications_monthly_by_workspace.pipe` file. The column is calculated using the `toStartOfMonth(fromUnixTimestamp64Milli(time))` function. This change allows for aggregating verifications monthly by workspace and includes the start of the month timestamp in the result.