Tinybird monthly pipes (#638)
* feat: add materialized views

* chore(mv__monthly_verifications.datasource): revert to VERSION 1 and update schema

The change was made to revert the datasource to VERSION 1 and update the schema. The previous version (VERSION 2) was created from a Pipe 'aggregate_verifications_daily__v2', but it was decided to revert to the previous version and update the schema accordingly.

* fix(aggregate_verifications_monthly_by_workspace.pipe): remove trailing comma in SQL query

The trailing comma in the SQL query was causing a syntax error. This commit removes the trailing comma to fix the issue.

* fix(aggregate_verifications_monthly_by_workspace.pipe): change GROUP BY clause to include workspaceId, apiId, keyId, and time

* feat(aggregate_verifications_monthly_by_workspace.pipe): add time column to the aggregation query

The change adds a new column called "time" to the aggregation query in the `aggregate_verifications_monthly_by_workspace.pipe` file. The column is calculated using the `toStartOfMonth(fromUnixTimestamp64Milli(time))` function. This change allows for aggregating verifications monthly by workspace and includes the start of the month timestamp in the result.