feat(v1_liveness.ts): add support for rate limiting and usage limiting services
This commit adds support for rate limiting and usage limiting services in the v1_liveness route. The `rateLimiter` and `usageLimiter` variables are imported from the `@/pkg/global` module and added to the list of imported variables.

In the `createRoute` function, two new properties `ratelimit` and `usagelimit` are added to the `config` object. These properties are of type `string` and have descriptions indicating that they represent the names of the connected rate limiting and usage limiting services, respectively.

In the `registerV1Liveness` function, the names of the rate limiting and usage limiting services are added to the `services` object that is returned in the response. The names are obtained from the corresponding variables `rateLimiter` and `usageLimiter` using the `constructor.name` property.