docs: Add Setup Instructions for Windows (#521)
* Add setup instructions for Windows

* chore(deploy-worker.yaml): remove deploy-worker.yaml workflow file

The deploy-worker.yaml workflow file was removed as it is no longer needed for the project.

feat(server): remove v1_key_verify.go file

The v1_key_verify.go file was removed from the server package. This file contained the implementation of the v1VerifyKey function, which handled the verification of API keys in the server. The function parsed the request body, validated the request, and performed various checks and validations before returning a response.

The removal of this file suggests that the v1VerifyKey functionality is no longer needed or has been refactored into another file or package.

---------

Co-authored-by: Andreas Thomas <dev@chronark.com>