fix: transaction deadlock in trpc
Fixes transaction deadlocks around audit log buckets.

This moves the workspace fetch to the beginning of the trpc handler and
loads the bucket id once, to be reused.

In theory we can now cache the workspace lookup but I'm not sure if
Next.js does that automatically now or if we need to do it.