fix: transaction deadlock in trpc (#2843)
* fix: transaction deadlock in trpc

Fixes transaction deadlocks around audit log buckets.

This moves the workspace fetch to the beginning of the trpc handler and
loads the bucket id once, to be reused.

In theory we can now cache the workspace lookup but I'm not sure if
Next.js does that automatically now or if we need to do it.

* fix: load the correct workspace id

* fix: race condition with deleting permissions

* fix: race condition with deleting roles

* fix: use correct id

* fix: remove duplicate query

* feat: add error log

* fix: populate workspace

* fix: populate workspace

* fix: remove opt in

* [autofix.ci] apply automated fixes

* fix: remove opt in

* revert: timeout

* revert: name

* revert: name

* revert: name

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>